DarkMode: Lower contrast of mouseover highlights
The mouseover highlights of the 3 main options as well as the menu options are geared more toward light mode. The contrast was too much for dark mode.

I kept the same DodgerBlue color and simply added a good amount of transparency so that it works well in both dark and light modes.

Also added a minor fix for Build page label background.