DarkMode: Lower contrast of mouseover highlights (#306)
* DarkMode: Lower contrast of text boxes

The text boxes were appearing with quite a jarring contrast in dark mode. I switched the text box BorderStyle to FixedSingle for a more visually appealing look with less contrast.

* DarkMode: Lower contrast of mouseover highlights

The mouseover highlights of the 3 main options as well as the menu options are geared more toward light mode. The contrast was too much for dark mode.

I kept the same DodgerBlue color and simply added a good amount of transparency so that it works well in both dark and light modes.

Also added a minor fix for Build page label background.