Set a standard design for Next and Reset buttons
This commit sets a standard design for Dark Mode and Light Mode for the Next button on the MainForm and also the Reset button on the Settings page. There is also a standard MouseOverBackColor that fits the colors from the rest of the app.

Once some of other issues with Dark/Light Button Controls (https://github.com/MicrosoftDocs/WDAC-Toolkit/issues/313) are fixed, I would like to expand so that this button color standard is applied to all other button throughout the app.