DarkMode: Fix radioButtons Dark/Light Color (#315)
The radiobuttons for File and Folder on the Exceptions page were hardcoded with a Black ForeColor. So the ForeColor was showing Black whether it was Dark or Light mode.

This fixes it by removing the hardcoded ForeColor, therefore allowing the already in place Dark/Light color control to work as expected.